+++ Final_Pr/src/main/java/by/bogdanov/dao/ClearanceDao.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface ClearanceDao extends BaseDao<Clearance>{
     List<Clearance> readAll() throws DaoException;
-    Clearance readById(Long id) throws DaoException;
-    void delete(Long id) throws DaoException;
+    Clearance readById(int id) throws DaoException;
+    void delete(int id) throws DaoException;
     void create(Clearance clearance) throws DaoException;
     void update(Clearance clearance) throws DaoException;
