+++ Final_Pr/src/main/java/by/bogdanov/dao/OperationDao.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface OperationDao extends BaseDao<Operation>{
     List<Operation> readAll() throws DaoException;
-    Operation readById(Long id) throws DaoException;
-    void delete(Long id) throws DaoException;
+    Operation readById(int id) throws DaoException;
+    void delete(int id) throws DaoException;
     void create(Operation operation) throws DaoException;
     void update(Operation operation) throws DaoException;
     Operation readByPrice(int price) throws DaoException;
