+++ Final_Pr/src/main/java/by/bogdanov/dao/OrderDao.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public interface OrderDao extends BaseDao<Order>{
     List<Order> readAll() throws DaoException;
-    Order readById(Long id) throws DaoException;
-    void delete(Long id) throws DaoException;
+    Order readById(int id) throws DaoException;
+    void delete(int id) throws DaoException;
     void create(Order order) throws DaoException;
     void update(Order order) throws DaoException;
-    List<Order> readByUserId(Long id) throws DaoException;
+    List<Order> readByUserId(int id) throws DaoException;
 }
