+++ Final_Pr/src/main/java/by/bogdanov/dao/VehicleDao.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface VehicleDao extends BaseDao<Vehicle> {
     List<Vehicle> readAll() throws DaoException;
-    Vehicle readById(Long id) throws DaoException;
-    void delete(Long id) throws DaoException;
+    Vehicle readById(int id) throws DaoException;
+    void delete(int id) throws DaoException;
     void create(Vehicle vehicle) throws DaoException;
     void update(Vehicle vehicle) throws DaoException;
     List<Vehicle> readByYear(int year) throws DaoException;
